adding a pending message component
